DELIA-57760: Issue with subtitle switching on LLAMA linear
Reason for change: Added a logic to retune
if the preferred track is not matching current
preferred values for both audio and text.
Risks: Medium
Test Procedure: Play Multi text content change the text
using json data dynamically, ensure json data have single values

(cherry picked from commit 566fbb6e368328afc16192ad8ee65b2312918023)

+++ fragmentcollector_mpd.h
@@ -366,7 +366,7 @@ public:
 	 * @param[out] selectedRepIdx - Selected representation index
 	 * @param[out] selectedRepBandwidth - selected audio track bandwidth
  	 */
-	void GetPreferredTextRepresentation(IAdaptationSet *adaptationSet, int &selectedRepIdx,	uint32_t &selectedRepBandwidth, uint64_t &score, std::string &name, std::string &codec);
+	void GetPreferredTextRepresentation(IAdaptationSet *adaptationSet, int &selectedRepIdx,	uint32_t &selectedRepBandwidth, unsigned long long &score, std::string &name, std::string &codec);
 	static Accessibility getAccessibilityNode(void *adaptationSet);
 	static Accessibility getAccessibilityNode(AampJsonObject &accessNode);
 	/**
